requirements.txt
click==7.1.2
pytest==6.2.1
pyyaml==5.3.1